Added InternalLink that consumes ref from material-ui (#77)
Function components in React may not be passed refs: (https://reactjs.org/docs/forwarding-refs.html#forwarding-refs-to-dom-components).

HashLink (RouterLink) is a function component. material-ui passes a ref down to the HashLink, which is not permitted. The new InternalLink component consumes the ref using React.forwardRef so it does not get passed to the HashLink anymore.